Allow parse to work on test1.myql
Rascal only applies `layout` within a syntax. This means that newlines
before and after the syntax are considered wrong. To allow us to
be able to directly call parse from ParseTree on the file, the newline
at the end of the file has been removed.

This does not appear to be an issue for the IDE. It likely targets just
the Form and ignores surrounding whitespace.

Signed-off-by: laburnumT <laburnumtec@gmail.com>
Signed-off-by: Oana <maimascu.oana@yahoo.com>